{"schema":"decodex/commit/1","summary":"Split CLI argument modules","authority":"manual"}
