+++ apps/elf-cli/src/args/constants.rs
@@ -1,0 +1,6 @@
+pub(in crate::args) const DEFAULT_API_URL: &str = "http://127.0.0.1:51892";
+pub(in crate::args) const DEFAULT_ADMIN_URL: &str = "http://127.0.0.1:51891";
+pub(in crate::args) const DEFAULT_TENANT_ID: &str = "local-tenant";
+pub(in crate::args) const DEFAULT_PROJECT_ID: &str = "local-project";
+pub(in crate::args) const DEFAULT_AGENT_ID: &str = "local-agent";
+pub(in crate::args) const DEFAULT_READ_PROFILE: &str = "private_only";
